Add files via upload
This is basic Java code. where the user can store the name of the user and his/her phone number. In the future, the phone number can be updated through the name of the user. The user can display all the phone numbers. or he/she can display the specific phone number.